Fix bug in sync protocol
The issue was that we rolled back the pending_sync watermark after any
error in the push job. Thus, if we were recovering from a failed push,
had a network failure, restarted the workload, and then pushed again we
would forget that we were recovering and push incorrect volume state to the
server.